LayoutTests/editing/pasteboard/clipboard-event-expected.txt

Tests the ClipboardEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firedCopyEvent is false
PASS firedPasteEvent is false
PASS event.type is "copy"
PASS event.__proto__ is ClipboardEvent.prototype
PASS event.isTrusted is true
PASS event.bubbles is true
PASS event.cancelable is true
PASS event.clipboardData.__proto__ is DataTransfer.prototype
PASS firedCopyEvent is true
PASS firedPasteEvent is false
PASS event.type is "paste"
PASS event.__proto__ is ClipboardEvent.prototype
PASS event.isTrusted is true
PASS event.bubbles is true
PASS event.cancelable is true
PASS event.clipboardData.__proto__ is DataTransfer.prototype
PASS event.clipboardData.getData('text/plain') is "Content"
PASS firedCopyEvent is false
PASS firedPasteEvent is true
PASS successfullyParsed is true

TEST COMPLETE
